create, edit & delete kata done

+++ src/domain/orm/kastas.orm.ts
@@ -1,6 +1,7 @@
 import { kataEntity } from "../entities/kata.entity";
-import { LogError, LogSucces } from "../../utils/logger";
+import { LogError, LogSucces, LogInfo } from '../../utils/logger';
 import { IKata } from "../interfaces/IKata.interfaces";
+import { userEntity } from "../entities/User.entity";
 
 // crud
 
@@ -46,41 +47,66 @@ export const getKatasByID = async (id: string): Promise<any | undefined> => {
     }
 }
 
-export const deleteKataByID = async (id: string): Promise<any | undefined> => {
+export const deleteKataByID = async (id: string, creatorId: string): Promise<any | undefined> => {
     try {
         let katasModel = kataEntity()
-        
-        return await katasModel.deleteOne({ _id: id })
+        let kata = await katasModel.findOne({ _id: id })
+
+        if (kata.creatorId === creatorId) {
+            return await katasModel.deleteOne({ _id: id })
+        } else {
+            throw new Error("Unauthorized to delete this kata")
+        }
     } catch (error) {
-        LogError(`[ORM ERROR]: Deleting katas by id ${error}`)
+        LogError(`[ORM ERROR]: Deleting katas by id \${error}`)
     }
 }
 
-export const getKatasPerDif = async (dif: number): Promise<any[] | undefined> => {
+export const getKatasPerDif = async (dif: string): Promise<any[] | undefined> => {
     try {
         let katasModel = kataEntity()
         //search
-        return await katasModel.find({ level: dif })
+        return await katasModel.find().sort({ level: -1 })
     } catch (error) {
         LogError(`[ORM ERROR]: Getting per dif katas ${error}`)
     }
 }
 
-export const createKata = async (kata: IKata): Promise<any[] | undefined> => {
+export const getKatasPerValoration = async (valoration: number): Promise<any[] | undefined> => {
     try {
         let katasModel = kataEntity()
+        //search
+        return await katasModel.find().sort({ valoration: -1 })
+    } catch (error) {
+        LogError(`[ORM ERROR]: Getting per dif katas ${error}`)
+    }
+}
+
+export const createKata = async (kata: IKata, idUser: string): Promise<any[] | undefined> => {
+    try {
+        let katasModel = kataEntity()
+        let userModel = userEntity()
+        const result = await katasModel.create(kata);
+        const kataId = result._id;
+
+        await userModel.updateOne({ _id: idUser }, { $push: { katas: kataId } });
         //create
-        return await katasModel.create(kata)
+        return await kataId
     } catch (error) {
         LogError(`[ORM ERROR]: Creating kata ${error}`)
     }
 }
 
-export const updateKataByid = async (id: string, kata: IKata): Promise<any | undefined> => {
+export const updateKataByid = async (id: string, kata: IKata,creatorId: string): Promise<any | undefined> => {
     try {
         let katasModel = kataEntity()
-        //update user
-        return await katasModel.findByIdAndUpdate(id, kata)
+        let kata = await katasModel.findOne({ _id: id })
+
+        if (kata.creatorId === creatorId) {
+            return await katasModel.findByIdAndUpdate(id, kata)
+        } else {
+            throw new Error("Unauthorized to delete this kata")
+        }
     } catch (error) {
         LogError(`[ORM ERROR]: Updating kata: ${id} ${error}`)
     }
@@ -140,9 +166,9 @@ export const updateValorationByID = async (id: string, kata: any, valoration: nu
         let katasModel = kataEntity()
         //search
         const kat = await katasModel.findById(id)
-        const nuevoPromedio = ((kat.valoration * kat.num_valoraciones) + valoration) / (kat.num_valoraciones + 1);
-        console.log(kat.valoration,kat.num_valoraciones,valoration)
-        return await katasModel.updateOne({ _id: id },{ $set: { valoration: nuevoPromedio.toFixed(1) },$inc: {num_valoraciones: 1}});
+        const nuevoPromedio = ((kat.valoration * kat.num_valorations) + valoration) / (kat.num_valorations + 1);
+        console.log(kat.valoration, kat.num_valorations, valoration)
+        return await katasModel.updateOne({ _id: id }, { $set: { valoration: nuevoPromedio.toFixed(1) }, $inc: { num_valorations: 1 } });
     } catch (error) {
         LogError(`[ORM ERROR]: Getting per valoration katas ${error}`)
     }
